+++ actions/bulk_apply_policies.py
@@ -5,7 +5,7 @@ from lib.actions import BaseAction
 
 class BulkApplyPolicies(BaseAction):
     """
-    Add/update address groups to a firewall in bulk
+    Add/update policies to a firewall in bulk
     """
     def run(self, class_string, device_group, firewall, objects, pre_rulebase):
 
@@ -32,8 +32,15 @@ class BulkApplyPolicies(BaseAction):
                     "{} contains invalid values for a {} object!".format(obj, cls.__name__)
                 )
 
-            pandevice_object = cls(**obj)
-            rulebase.add(pandevice_object)
+            # manually update an existing object or add a new one
+            pandevice_object = rulebase.find(obj['name'], class_type=cls)
+            if pandevice_object is not None:
+                for key, value in obj.items():
+                    setattr(pandevice_object, key, value)
+
+            else:
+                pandevice_object = cls(**obj)
+                rulebase.add(pandevice_object)
 
         pandevice_object.apply_similar()